Move gastro to apps folder
