+++ src/apps/Gastro/components/AreaMap.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import MapboxGL from 'mapbox-gl';
 import styled from 'styled-components';
 import Map from '~/components2/Map';
-import config from '~/pages/Gastro/config';
+import config from '~/apps/Gastro/config';
 
 const StyledMap = styled(Map)`
   width: 100%;
