+++ src/apps/Gastro/components/Notice.tsx
