+++ src/apps/Gastro/components/RegistrationForm/SectionArea.tsx
