+++ src/apps/Gastro/components/RegistrationForm/SectionBase.tsx
