+++ src/apps/Gastro/components/RegistrationForm/SectionCertificate.tsx
