+++ src/apps/Gastro/components/RegistrationForm/SectionEmail.tsx
