+++ src/apps/Gastro/components/RegistrationForm/SectionNotice.tsx
