+++ src/apps/Gastro/components/RegistrationForm/validate.ts
