+++ src/apps/Gastro/components/Thanks.tsx
