+++ src/apps/Gastro/config.ts
