+++ src/apps/Gastro/pages/Signup.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import { Container } from '@material-ui/core';
 import styled from 'styled-components';
 
-import config from '~/pages/Gastro/config';
+import config from '~/apps/Gastro/config';
 import Header from '../components/Header';
 import Thanks from '../components/Thanks';
 import SignupForm from '../components/SignupForm';
