+++ src/apps/Gastro/regulations.ts
