+++ src/components2/Map/Map.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useRef, useEffect } from 'react';
 import MapboxGL from 'mapbox-gl';
 import styled from 'styled-components';
-import config from '~/pages/Gastro/config';
+import config from '~/apps/Gastro/config';
 import logger from '~/utils/logger';
 
 const Wrapper = styled.div`
